// resources/js/api/generated/models/sessionItem.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface SessionItem {
  bot: boolean;
  browser: string;
  created_at: string;
  current: boolean;
  desktop: boolean;
  device: string;
  device_type: string;
  expires_at: string;
  /** @minimum 0 */
  id: number;
  ip_address: string;
  last_used: string;
  location: string;
  mobile: boolean;
  os: string;
  tablet: boolean;
  token: string;
  type: string;
  user_agent: string;
  /** @minimum 0 */
  user_id: number;
}